+++ package.json
@@ -45,6 +45,7 @@
     "@rollup/plugin-commonjs": "^21.0.1",
     "@rollup/plugin-node-resolve": "^13.1.3",
     "@rollup/plugin-url": "^6.1.0",
+    "@types/react-table": "^7.7.11",
     "clsx": "1.1.1",
     "create-react-app": "^5.0.0",
     "immutability-helper": "^3.1.1",
@@ -56,6 +57,7 @@
     "react-table": "7.7.0",
     "react-window": "^1.8.6",
     "rollup": "^2.67.2",
+    "rollup-plugin-esbuild": "^4.9.1",
     "rollup-plugin-peer-deps-external": "^2.2.4",
     "rollup-plugin-postcss": "^4.0.2",
     "styled-components": "^5.3.5"
