+++ rollup.config.js
@@ -1,13 +1,9 @@
-import babel from '@rollup/plugin-babel'
-import commonjs from '@rollup/plugin-commonjs'
-import external from 'rollup-plugin-peer-deps-external'
-import resolve from '@rollup/plugin-node-resolve'
-import url from '@rollup/plugin-url'
+import esbuild from 'rollup-plugin-esbuild'
 
 import pkg from './package.json'
 
 export default {
-  input: 'src/index.js',
+  input: 'src/index.ts',
   output: [
     {
       file: pkg.main,
@@ -20,14 +16,31 @@ export default {
       sourcemap: true
     }
   ],
+  
   plugins: [
-    external(),
-    url(),
-    babel({
-      exclude: 'node_modules/**',
-      babelHelpers: 'bundled',
+    esbuild({
+      // All options are optional
+      include: /\.[jt]sx?$/, // default, inferred from `loaders` option
+      exclude: /node_modules/, // default
+      sourceMap: false, // by default inferred from rollup's `output.sourcemap` option
+      minify: process.env.NODE_ENV === 'production',
+      target: 'es2017', // default, or 'es20XX', 'esnext'
+      jsx: 'transform', // default, or 'preserve'
+      jsxFactory: 'React.createElement',
+      jsxFragment: 'React.Fragment',
+      // Like @rollup/plugin-replace
+      define: {
+        __VERSION__: '"x.y.z"',
+      },
+      tsconfig: 'tsconfig.json', // default
+      // Add extra loaders
+      loaders: {
+        // Add .json files support
+        // require @rollup/plugin-commonjs
+        '.json': 'json',
+        // Enable JSX in .js files too
+        '.js': 'jsx',
+      },
     }),
-    resolve(),
-    commonjs()
   ]
 }
