+++ src/Cell.tsx
@@ -1,9 +1,9 @@
 import React, { useEffect, useState } from 'react';
 import ContentEditable from 'react-contenteditable';
-import Badge from './Badge.jsx';
+import Badge from './Badge';
 import { usePopper } from 'react-popper';
 import { grey } from './colors';
-import PlusIcon from './img/Plus.jsx';
+import PlusIcon from './img/Plus';
 import { ActionTypes, DataTypes, randomColor } from './utils';
 import { createPortal } from 'react-dom';
