+++ src/img/Hash.tsx
