+++ src/img/Text.tsx
