+++ src/img/Trash.tsx
