--- src/index.js
@@ -1,5 +1,0 @@
-export { default as Table } from './Table.jsx';
-export { default as TableWrapper } from './TableWrapper.jsx';
-export { default as utils } from './utils';
-export { default as Cell } from './Cell.jsx';
-export { default as Badge } from './Badge.jsx';
