+++ src/index.ts
@@ -1,0 +1,5 @@
+export { default as Table } from './Table';
+export { default as TableWrapper } from './TableWrapper';
+export { default as utils } from './utils';
+export { default as Cell } from './Cell';
+export { default as Badge } from './Badge';
