+++ src/scrollbarWidth.ts
